Refactor docs header and fix code editor styles
