--- docs/src/components/header.mdx
@@ -1,10 +1,0 @@
-import pkg from '@mdx-js/mdx/package.json'
-import Logo from './logo'
-
-<a href='/'>
-  <Logo /> MDX v{pkg.version}
-</a>
-
-<div style={{ margin: 'auto' }} />
-
-[![GitHub](https://icon.now.sh/github/24)](https://github.com/mdx-js/mdx 'GitHub')
