--- docs/src/components/logo.js
@@ -1,5 +1,0 @@
-import React from 'react'
-
-export default () => (
-  <img src="https://mdx-logo.now.sh" width="70" alt="MDX logo" />
-)
